requirements.txt
fastapi==0.115
geopy==2.4.1
npindex==0.1.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pillow_heif==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.32.3
rich==13.9.4
selenium==4.26.1
uvicorn==0.32.0